get rid of DevicePatchMixin
