+++ patches/patch.py
@@ -269,43 +269,3 @@ class FirmwarePatchMixin:
                 raise KeyError(f"0x{val:08X} at offset 0x{offset:08X}")
             self[offset : offset + size] = new_val.to_bytes(size, "little")
             # print(f"    lookup 0x{val:08X}->0x{new_val:08x}")
-
-
-class DevicePatchMixin:
-    def _move_copy(
-        self, dst, dst_offset: int, src, src_offset: int, size: int, delete: bool
-    ) -> int:
-        dst[dst_offset : dst_offset + size] = src[src_offset : src_offset + size]
-        if delete:
-            src.clear_range(src_offset, src_offset + size)
-
-        for i in range(size):
-            self.lookup[src.FLASH_BASE + src_offset + i] = (
-                dst.FLASH_BASE + dst_offset + i
-            )
-
-        return size
-
-    def _move(self, dst, dst_offset: int, src, src_offset: int, size: int) -> int:
-        return self._move_copy(dst, dst_offset, src, src_offset, size, True)
-
-    def _copy(self, dst, dst_offset: int, src, src_offset: int, size: int) -> int:
-        return self._move_copy(dst, dst_offset, src, src_offset, size, False)
-
-    # Convenience methods for move and copy
-    def _move_ext_to_int(self, ext_offset: int, int_offset: int, size: int) -> int:
-        return self._move(self.internal, int_offset, self.external, ext_offset, size)
-
-    def _copy_ext_to_int(self, ext_offset: int, int_offset: int, size: int) -> int:
-        return self._copy(self.internal, int_offset, self.external, ext_offset, size)
-
-    def _move_to_compressed_memory(
-        self, ext_offset: int, compressed_memory_offset: int, size: int
-    ) -> int:
-        return self._move(
-            self.compressed_memory,
-            compressed_memory_offset,
-            self.external,
-            ext_offset,
-            size,
-        )
